Moved DefaultLanguageCodeForContentTest to Integration scope
