+++ phpunit-integration-legacy.xml
@@ -74,8 +74,8 @@
         <testsuite name="eZ\Publish\Core\FieldType\Tests\Integration">
             <directory>eZ/Publish/Core/FieldType/Tests/Integration</directory>
         </testsuite>
-        <testsuite name="Ibexa\Tests\Core\Repository\Values\Content">
-            <file>tests/lib/Repository/Values/Content/DefaultLanguageCodeForContentTest.php</file>
+        <testsuite name="Ibexa\Tests\Integration\Core\Repository\Values\Content">
+            <file>tests/integration/Core/Repository/Values/Content/DefaultLanguageCodeForContentTest.php</file>
         </testsuite>
         <testsuite name="gateway-integration">
             <file>eZ/Publish/SPI/Tests/Search/Content/IndexerGatewayTest.php</file>
